% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_rate}
\alias{as_rate}
\title{Rate conversion (decimal vs bps)}
\usage{
as_rate(dec = NULL, bps = NULL)
}
\arguments{
\item{dec}{numeric(1). Decimal rate.}

\item{bps}{numeric(1). Basis points.}
}
\value{
numeric(1) as decimal.
}
\description{
Rate conversion (decimal vs bps)
}
