% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cfg_validate}
\alias{cfg_validate}
\title{Validate YAML configuration structure}
\usage{
cfg_validate(cfg)
}
\arguments{
\item{cfg}{list returned by dcf_read_config().}
}
\value{
cfg invisibly (or error if invalid).
}
\description{
Validate YAML configuration structure
}
