% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf.R
\name{compute_noi_y1}
\alias{compute_noi_y1}
\title{Quick computation of year-1 NOI}
\usage{
compute_noi_y1(rent_signed, lettable_area, vac_rate = 0)
}
\arguments{
\item{rent_signed}{numeric(1). Face rent (€/m²/year).}

\item{lettable_area}{numeric(1). Lettable area (m²).}

\item{vac_rate}{numeric(1) in \code{[0,1)}. Average vacancy rate.}
}
\value{
numeric(1) \eqn{NOI_{y1}} rounded to cents.
}
\description{
Quick computation of year-1 NOI
}
\examples{
compute_noi_y1(400, 2500, vac_rate = 0.05)
}
