% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{dcf_read_config}
\alias{dcf_read_config}
\title{Read a configuration YAML}
\usage{
dcf_read_config(
  config_file = system.file("extdata", "preset_default.yml", package = "cre.dcf")
)
}
\arguments{
\item{config_file}{path; default to inst/extdata/config.yml in the package.}
}
\value{
list
}
\description{
Read a configuration YAML
}
