% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dcf_write_yaml_template}
\alias{dcf_write_yaml_template}
\title{Write a commented YAML template for users to edit}
\usage{
dcf_write_yaml_template(path)
}
\arguments{
\item{path}{File path where to write the \code{'YAML'} file
(for example \code{"my_case.yml"}).}
}
\value{
The input \code{path}, invisibly.
}
\description{
Creates a \code{'YAML'} file on disk from \code{dcf_spec_template()},
suitable for manual editing.
}
\examples{
tmp <- tempfile(fileext = ".yml")
dcf_write_yaml_template(tmp)
stopifnot(file.exists(tmp))
unlink(tmp)

}
