% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_cfg}
\alias{get_cfg}
\title{Safe access to nested YAML values}
\usage{
get_cfg(cfg, ..., default = NULL)
}
\arguments{
\item{cfg}{list configuration object.}

\item{...}{nested keys.}

\item{default}{value if missing.}
}
\value{
value or default.
}
\description{
Safe access to nested YAML values
}
