% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guard_rate}
\alias{guard_rate}
\title{Guardrail on an input rate (message if scale likely incorrect)}
\usage{
guard_rate(x, name)
}
\arguments{
\item{x}{numeric(1).}

\item{name}{character(1). Parameter name used in messages.}
}
\value{
numeric(1) unchanged.
}
\description{
Guardrail on an input rate (message if scale likely incorrect)
}
