% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.R
\name{interest_rule}
\alias{interest_rule}
\title{Build an interest-deductibility rule for the generic SPV tax engine}
\usage{
interest_rule(mode = c("full"))
}
\arguments{
\item{mode}{Character scalar. Currently only \code{"full"} is supported.}
}
\value{
A list of class \code{cre_tax_interest_rule}.
}
\description{
Build an interest-deductibility rule for the generic SPV tax engine
}
\examples{
interest_rule()
}
