% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_api.R
\name{lease_roll}
\alias{lease_roll}
\title{Group lease units into a simplified lease roll}
\usage{
lease_roll(units)
}
\arguments{
\item{units}{List of objects returned by \code{\link[=lease_unit]{lease_unit()}}.}
}
\value{
An object of class \code{cre_lease_roll}.
}
\description{
Group lease units into a simplified lease roll
}
\examples{
roll <- lease_roll(list(
  lease_unit(
    "North",
    area_sqm = 1200,
    events = list(lease_event(start = 2025, end = 2027, rent = 220))
  )
))
length(roll$units)
}
