% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_api.R
\name{lease_roll_snapshot}
\alias{lease_roll_snapshot}
\title{Summarize a lease roll in analyst-friendly tabular form}
\usage{
lease_roll_snapshot(x)
}
\arguments{
\item{x}{Object returned by \code{\link[=lease_unit]{lease_unit()}} or \code{\link[=lease_roll]{lease_roll()}}.}
}
\value{
A tibble with one row per lease unit.
}
\description{
Summarize a lease roll in analyst-friendly tabular form
}
