% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.R
\name{loss_rule}
\alias{loss_rule}
\title{Build a loss-carryforward rule for the generic SPV tax engine}
\usage{
loss_rule(carryforward = TRUE, carryforward_years = Inf, offset_cap_pct = 1)
}
\arguments{
\item{carryforward}{Logical scalar. Whether tax losses can be carried
forward.}

\item{carryforward_years}{Numeric scalar. Number of future years for which a
loss may be used. Use \code{Inf} for no expiry.}

\item{offset_cap_pct}{Numeric scalar in \code{[0, 1]}. Maximum fraction of a
positive taxable base that can be offset by prior losses.}
}
\value{
A list of class \code{cre_tax_loss_rule}.
}
\description{
Build a loss-carryforward rule for the generic SPV tax engine
}
\examples{
loss_rule(carryforward = TRUE, carryforward_years = Inf, offset_cap_pct = 1)
}
