% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{run_from_config}
\alias{run_from_config}
\title{Canonical pipeline from a YAML file}
\usage{
run_from_config(config_file, ltv_base = c("price_ht", "price_di", "value"))
}
\arguments{
\item{config_file}{path to YAML.}

\item{ltv_base}{"price_ht" | "price_di" | "value".}
}
\value{
list(dcf, debt, full, ratios, norm)
}
\description{
Canonical pipeline from a YAML file
}
