% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.R
\name{tax_spec_spv}
\alias{tax_spec_spv}
\title{Build a generic SPV tax specification}
\usage{
tax_spec_spv(
  corp_tax_rate = 0.25,
  depreciation_spec = NULL,
  interest_rule = NULL,
  loss_rule = NULL
)
}
\arguments{
\item{corp_tax_rate}{Numeric scalar in \code{[0, 1)}.}

\item{depreciation_spec}{Object returned by \code{\link[=depreciation_spec]{depreciation_spec()}}.}

\item{interest_rule}{Object returned by \code{\link[=interest_rule]{interest_rule()}}.}

\item{loss_rule}{Object returned by \code{\link[=loss_rule]{loss_rule()}}.}
}
\value{
A list of class \code{cre_tax_spec_spv}.
}
\description{
Build a generic SPV tax specification
}
\examples{
spec <- tax_spec_spv(corp_tax_rate = 0.25)
spec$corp_tax_rate
}
