% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_style_preset}
\alias{load_style_preset}
\title{Load a preset style YAML file}
\usage{
load_style_preset(style)
}
\description{
This loader reads preset YAML files from \code{inst/extdata} inside the installed
package.
}
\keyword{internal}
